Package name Refactor
This includes changes to rename package to
"com.intel.sample.dms" as the object detection and dms
both where having same package name
