+++ android/app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdkVersion 28
     buildToolsVersion '28.0.3'
     defaultConfig {
-        applicationId "org.tensorflow.lite.examples.detection"
+        applicationId "com.intel.sample.dms"
         minSdkVersion 21
         targetSdkVersion 21
         versionCode 1
