+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.tensorflow.lite.examples.detection">
+    package="com.intel.sample.dms">
 
     <uses-sdk />
 
@@ -18,7 +18,7 @@
         android:theme="@style/AppTheme.ObjectDetection">
 
         <activity
-            android:name=".DetectorActivity"
+            android:name="com.intel.sample.dms.DetectorActivity"
             android:label="@string/tfe_od_app_name" >
 <!--            android:screenOrientation="portrait">-->
             <intent-filter>
