+++ android/app/src/main/java/com/intel/sample/dms/ImageUtils.java
@@ -1,0 +1,157 @@
+/* Copyright 2015 The TensorFlow Authors. All Rights Reserved.
+   Copyright (C) 2022-2023 Intel Corporation
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+SPDX-License-Identifier: Apache-2.0
+==============================================================================*/
+
+package com.intel.sample.dms;
+
+import android.graphics.Bitmap;
+import android.graphics.Canvas;
+import android.graphics.Matrix;
+import android.os.Environment;
+import android.util.Log;
+
+import java.io.File;
+import java.io.FileOutputStream;
+
+/**
+ * Utility class for manipulating images.
+ **/
+public class ImageUtils {
+    private static final String TAG = "ImageUtils";
+    /**
+     * Saves a Bitmap object to disk for analysis.
+     *
+     * @param bitmap The bitmap to save.
+     */
+    public static void saveBitmap(final Bitmap bitmap) {
+        saveBitmap(bitmap, "preview.png");
+    }
+
+    /**
+     * Saves a Bitmap object to disk for analysis.
+     *
+     * @param bitmap   The bitmap to save.
+     * @param filename The location to save the bitmap to.
+     */
+    public static void saveBitmap(final Bitmap bitmap, final String filename) {
+        final String root =
+                Environment.getExternalStorageDirectory().getAbsolutePath() + File.separator + "tensorflow";
+        Log.i(TAG, String.format("Saving %dx%d bitmap to %s.", bitmap.getWidth(), bitmap.getHeight(), root));
+        final File myDir = new File(root);
+
+        if (!myDir.mkdirs()) {
+            Log.i(TAG, "Make dir failed");
+        }
+
+        final String fname = filename;
+        final File file = new File(myDir, fname);
+        if (file.exists()) {
+            try {
+              file.delete();
+	    } catch (Exception e) {
+              Log.i(TAG, "saveBitmap: failed to delete file: " + filename);
+	    }
+        }
+        try {
+            final FileOutputStream out = new FileOutputStream(file);
+            bitmap.compress(Bitmap.CompressFormat.PNG, 99, out);
+            out.flush();
+            out.close();
+        } catch (final Exception e) {
+            Log.e(TAG, "Exception!" + e);
+        }
+    }
+
+    /**
+     * Returns a transformation matrix from one reference frame into another.
+     * Handles cropping (if maintaining aspect ratio is desired) and rotation.
+     *
+     * @param srcWidth            Width of source frame.
+     * @param srcHeight           Height of source frame.
+     * @param dstWidth            Width of destination frame.
+     * @param dstHeight           Height of destination frame.
+     * @param applyRotation       Amount of rotation to apply from one frame to another.
+     *                            Must be a multiple of 90.
+     * @param maintainAspectRatio If true, will ensure that scaling in x and y remains constant,
+     *                            cropping the image if necessary.
+     * @return The transformation fulfilling the desired requirements.
+     */
+    public static Matrix getTransformationMatrix(
+            final int srcWidth,
+            final int srcHeight,
+            final int dstWidth,
+            final int dstHeight,
+            final int applyRotation,
+            final boolean maintainAspectRatio) {
+        final Matrix matrix = new Matrix();
+
+        if (applyRotation != 0) {
+            if (applyRotation % 90 != 0) {
+            }
+
+            // Translate so center of image is at origin.
+            matrix.postTranslate(-srcWidth / 2.0f, -srcHeight / 2.0f);
+
+            // Rotate around origin.
+            matrix.postRotate(applyRotation);
+        }
+
+        // Account for the already applied rotation, if any, and then determine how
+        // much scaling is needed for each axis.
+        final boolean transpose = (Math.abs(applyRotation) + 90) % 180 == 0;
+
+        final int inWidth = transpose ? srcHeight : srcWidth;
+        final int inHeight = transpose ? srcWidth : srcHeight;
+
+        // Apply scaling if necessary.
+        if (inWidth != dstWidth || inHeight != dstHeight) {
+            final float scaleFactorX = dstWidth / (float) inWidth;
+            final float scaleFactorY = dstHeight / (float) inHeight;
+
+            if (maintainAspectRatio) {
+                // Scale by minimum factor so that dst is filled completely while
+                // maintaining the aspect ratio. Some image may fall off the edge.
+                final float scaleFactor = Math.max(scaleFactorX, scaleFactorY);
+                matrix.postScale(scaleFactor, scaleFactor);
+            } else {
+                // Scale exactly to fill dst from src.
+                matrix.postScale(scaleFactorX, scaleFactorY);
+            }
+        }
+
+        if (applyRotation != 0) {
+            // Translate back from origin centered reference to destination frame.
+            matrix.postTranslate(dstWidth / 2.0f, dstHeight / 2.0f);
+        }
+
+        return matrix;
+    }
+
+    public static Bitmap resize(Bitmap input, int width, int height, int orientation) {
+        Bitmap bitmap = Bitmap.createBitmap(width, height, Bitmap.Config.ARGB_8888);
+        final Canvas canvas = new Canvas(bitmap);
+        Matrix frameToCropTransform = getTransformationMatrix(
+                input.getWidth(), input.getHeight(),
+                width, height,
+                orientation, false);
+//        Matrix cropToFrameTransform = new Matrix();
+//        frameToCropTransform.invert(cropToFrameTransform);
+        canvas.drawBitmap(input, frameToCropTransform, null);
+//        saveBitmap(bitmap);
+        return bitmap;
+    }
+}
