+++ android/app/src/main/java/com/intel/sample/dms/customview/ResultsView.java
@@ -1,0 +1,23 @@
+/* Copyright 2019 The TensorFlow Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+
+package com.intel.sample.dms.customview;
+
+import java.util.List;
+import com.intel.sample.dms.detection.tflite.Detector.Recognition;
+
+public interface ResultsView {
+  public void setResults(final List<Recognition> results);
+}
