+++ android/app/src/main/res/layout/tfe_od_camera_connection_fragment_tracking.xml
@@ -17,12 +17,12 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <org.tensorflow.lite.examples.detection.customview.AutoFitTextureView
+    <com.intel.sample.dms.customview.AutoFitTextureView
         android:id="@+id/texture"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content" />
 
-    <org.tensorflow.lite.examples.detection.customview.OverlayView
+    <com.intel.sample.dms.customview.OverlayView
         android:id="@+id/tracking_overlay"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
