+++ android/lib_interpreter/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.tensorflow.lite.examples.detection.tflite">
+    package="com.intel.sample.dms">
 </manifest>
